+++ code/game/objects/structures/reagent_dispensers.dm
@@ -46,9 +46,17 @@
 	create_reagents(tank_volume, AMOUNT_VISIBLE|DRAINABLE, list_reagents)
 
 /obj/structure/reagent_dispensers/ex_act(severity)
-	if(prob(severity / 4))
-		new /obj/effect/particle_effect/water(loc)
-		qdel(src)
+	switch(severity)
+		if(EXPLODE_DEVASTATE)
+			qdel(src)
+		if(EXPLODE_HEAVY)
+			if (prob(50))
+				new /obj/effect/particle_effect/water(loc)
+				qdel(src)
+		if(EXPLODE_LIGHT)
+			if (prob(5))
+				new /obj/effect/particle_effect/water(loc)
+				qdel(src)
 
 //Dispensers
 /obj/structure/reagent_dispensers/watertank
@@ -59,9 +67,7 @@
 	amount_per_transfer_from_this = 10
 	list_reagents = list(/datum/reagent/water = 1000)
 
-/obj/structure/reagent_dispensers/watertank/pred
-	icon = 'icons/obj/machines/yautja_machines.dmi'
-	icon_state = "watertank"
+
 
 /obj/structure/reagent_dispensers/fueltank
 	name = "fueltank"
@@ -75,10 +81,6 @@
 	var/obj/item/assembly_holder/rig
 	//Whether the tank is already exploding to prevent chain explosions
 	var/exploding = FALSE
-
-/obj/structure/reagent_dispensers/fueltank/pred
-	icon = 'icons/obj/machines/yautja_machines.dmi'
-	icon_state = "weldtank"
 
 /obj/structure/reagent_dispensers/fueltank/Destroy()
 	QDEL_NULL(rig)
